Review Feedack added

+++ src/batcontrol/logic/common.py
@@ -92,11 +92,11 @@ class CommonLogic:
 
         if capacity >= self.max_capacity * self.always_allow_discharge_limit:
             logger.debug(
-                'Discharge is \'always allowed\' for current capacity: %d Wh', round(capacity,0))
+                'Discharge is \'always allowed\' for current capacity: %.0f Wh', round(capacity,0))
             return True
 
         logger.debug(
-            'Discharge is NOT \'always allowed\' for current capacity: %d Wh', round(capacity,0))
+            'Discharge is NOT \'always allowed\' for current capacity: %.0f Wh', round(capacity,0))
         return False
 
     def is_charging_above_minimum(self, needed_energy: float) -> bool:
@@ -109,7 +109,7 @@ class CommonLogic:
             return True
 
         logger.debug(
-            'Charging needed recharge energy is below threshold(%d): %d Wh',
+            'Charging needed recharge energy is below threshold(%.0f): %.0f Wh',
                      round(self.min_charge_energy,0),
                      round(needed_energy,0))
         return False
@@ -119,7 +119,7 @@ class CommonLogic:
         Args:
             charge_rate (float): The initial charge rate in W.
         Returns:
-            float: The adjusted charge rate in W."""
+            int: The adjusted charge rate in W."""
         logger.debug('Calculating charge rate: %s', charge_rate)
         adjusted_charge_rate = charge_rate * self.charge_rate_multiplier
         if adjusted_charge_rate < MIN_CHARGE_RATE:
@@ -129,5 +129,5 @@ class CommonLogic:
             adjusted_charge_rate = MIN_CHARGE_RATE
 
         adjusted_charge_rate = int(round(adjusted_charge_rate, 0))
-        logger.debug('Adjusted charge rate: %.1f W', adjusted_charge_rate)
+        logger.debug('Adjusted charge rate: %d W', adjusted_charge_rate)
         return adjusted_charge_rate
